chore: Update Dockerfile to install gcc-riscv64-unknown-elf package
The Dockerfile in the .devcontainer directory is updated to include the installation of the `gcc-riscv64-unknown-elf` package. This change ensures that the necessary tool is available for building RISC-V code within the development container.

The purpose of this change is to provide support for compiling RISC-V code using the `gcc-riscv64-unknown-elf` compiler.

The commit also includes an update to the Makefile to add a new `run` target, which runs the compiled RISC-V code using QEMU.